requirements.txt
streamlit
altair
gensim
pandas
numpy
scikit-learn
umap-learn
pacmap
matplotlib
